third_party/WebKit/LayoutTests/animations/svg/css-animation-overrides-svg-presentation-attribute-animation.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<svg>
  <rect id="target"/>
</svg>

<script>
test(() => {
  target.animate({color: 'green'}, {fill: 'forwards'});
  target.animate({'svg-color': 'red'}, {fill: 'forwards'});
  assert_equals(getComputedStyle(target).color, 'rgb(0, 128, 0)');
}, 'CSS animations always override SVG presentation attribute animations');
</script>
